# Document/swagger-insure.v1.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "RFQ"
    },
    "servers": [
        {
            "url": "http://192.168.3.56:9763",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/rfq/expiryDetails/updateExpiryDetails/{id}": {
            "put": {
                "tags": [
                    "expiry-details-controller"
                ],
                "operationId": "updateExpiryDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExpiryDetailsDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExpiryPolicyDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/clientList/updateClientUserById/{uid}": {
            "put": {
                "tags": [
                    "client-list-user-controller"
                ],
                "operationId": "updateClientUserById",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClientListUserDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClientListUserDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/roles/create-userLogin": {
            "post": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "createUser",
                "parameters": [
                    {
                        "name": "locationId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "designationId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "departmentId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UsersNewDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/UsersNewDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/rfq/header/addHeader": {
            "post": {
                "tags": [
                    "excel-report-headers-controller"
                ],
                "operationId": "addHeader",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExcelReportHeadersDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExcelReportHeadersDto"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rfq/expiryDetails/createExpiryDetails": {
            "post": {
                "tags": [
                    "expiry-details-controller"
                ],
                "operationId": "createExpiryDetails",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ExpiryDetailsDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/products/addProduct/{clientListId}": {
            "post": {
                "tags": [
                    "product-controller"
                ],
                "operationId": "createClientListUser",
                "parameters": [
                    {
                        "name": "clientListId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddProductClientDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AddProductClientDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/clientList/addUser/{id}": {
            "post": {
                "tags": [
                    "client-list-user-controller"
                ],
                "operationId": "createClientListUser_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddClientListUserDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AddClientListUserDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/roles/getAllUsersByLocationIdLogin/{locationId}": {
            "get": {
                "tags": [
                    "user-controller"
                ],
                "operationId": "getAllUsers",
                "parameters": [
                    {
                        "name": "locationId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UsersNewDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/rfq/expiryDetails/getExpiryDetailsById": {
            "get": {
                "tags": [
                    "expiry-details-controller"
                ],
                "operationId": "getExpiryDetailsById",
                "parameters": [
                    {
                        "name": "rfqId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExpiryPolicyDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/products/getAllproduct/{clientListId}": {
            "get": {
                "tags": [
                    "product-controller"
                ],
                "operationId": "getAllUsers_1",
                "parameters": [
                    {
                        "name": "clientListId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductClientDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/clientList/getAllUsers/{id}": {
            "get": {
                "tags": [
                    "client-list-user-controller"
                ],
                "operationId": "getAllUsers_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ClientListUserDto"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/clientList/deleteUserByClientListId/{id}": {
            "delete": {
                "tags": [
                    "client-list-user-controller"
                ],
                "operationId": "deleteUserByClientListId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ExpiryDetailsDto": {
                "type": "object",
                "properties": {
                    "rfqId": {
                        "type": "string"
                    },
                    "policyNumber": {
                        "type": "string"
                    },
                    "startPeriod": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "endPeriod": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "premiumPaidInception": {
                        "type": "number",
                        "format": "double"
                    },
                    "premium": {
                        "type": "number",
                        "format": "double"
                    },
                    "additionPremium": {
                        "type": "number",
                        "format": "double"
                    },
                    "deletionPremium": {
                        "type": "number",
                        "format": "double"
                    },
                    "policyType": {
                        "type": "string"
                    },
                    "activeYears": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "membersNoInception": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "additions": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "deletions": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalMembers": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "membersNum": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "dependentMember": {
                        "type": "string"
                    },
                    "familyDefination": {
                        "type": "string"
                    },
                    "familiesNum": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "additionalRelationShip": {
                        "type": "string"
                    },
                    "familyDefinationRevision": {
                        "type": "string"
                    },
                    "createDate": {
                        "type": "string"
                    },
                    "updateDate": {
                        "type": "string"
                    },
                    "recordStatus": {
                        "type": "string"
                    }
                }
            },
            "ExpiryPolicyDetails": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "rfqId": {
                        "type": "string"
                    },
                    "policyNumber": {
                        "type": "string"
                    },
                    "startPeriod": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "endPeriod": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "premiumPaidInception": {
                        "type": "number",
                        "format": "double"
                    },
                    "premium": {
                        "type": "number",
                        "format": "double"
                    },
                    "additionPremium": {
                        "type": "number",
                        "format": "double"
                    },
                    "deletionPremium": {
                        "type": "number",
                        "format": "double"
                    },
                    "policyType": {
                        "type": "string"
                    },
                    "activeYears": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "membersNoInception": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "additions": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "deletions": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "totalMembers": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "membersNum": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "dependentMember": {
                        "type": "string"
                    },
                    "familyDefination": {
                        "type": "string"
                    },
                    "familiesNum": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "additionalRelationShip": {
                        "type": "string"
                    },
                    "familyDefinationRevision": {
                        "type": "string"
                    },
                    "createDate": {
                        "type": "string"
                    },
                    "updateDate": {
                        "type": "string"
                    },
                    "recordStatus": {
                        "type": "string"
                    }
                }
            },
...
            "ClientListUserDto": {
                "type": "object",
                "properties": {
                    "uid": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "employeeId": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "designation": {
                        "type": "string"
                    },
                    "mailId": {
                        "type": "string"
                    },
                    "phoneNo": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "createdDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedDate": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "UsersNewDto": {
                "required": [
                    "age",
                    "businessType",
                    "corporateName",
                    "dateOfBirth",
                    "email",
                    "employeeId",
                    "firstName",
                    "gender",
                    "lastName",
                    "password",
                    "phoneNo"
                ],
                "type": "object",
                "properties": {
                    "corporateName": {
                        "type": "string"
                    },
                    "businessType": {
                        "type": "string"
                    },
                    "firstName": {
                        "type": "string"
                    },
                    "lastName": {
                        "type": "string"
                    },
                    "employeeId": {
                        "type": "string"
                    },
                    "dateOfBirth": {
                        "type": "string",
                        "format": "date"
                    },
                    "age": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "gender": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phoneNo": {
                        "maxLength": 10,
                        "minLength": 10,
                        "pattern": "^[0-9]+$",
                        "type": "string"
                    },
                    "password": {
                        "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@#$%!^&*()])[A-Za-z\\d@#$%!^&*()]{5,11}$",
                        "type": "string"
                    }
                }
            },
...
            "ExcelReportHeadersDto": {
                "type": "object",
                "properties": {
                    "headerName": {
                        "type": "string"
                    },
                    "headerCategory": {
                        "type": "string"
                    },
                    "headerAliasname": {
                        "$ref": "#/components/schemas/ExcelReportHeadersMappingDto"
                    },
                    "CREATEDDATE": {
                        "type": "string"
                    },
                    "UPDATEDDATE": {
                        "type": "string"
                    },
                    "STATUS": {
                        "type": "string"
                    }
                }
            },
            "ExcelReportHeadersMappingDto": {
                "type": "object",
                "properties": {
                    "aliasName": {
                        "type": "string"
                    },
                    "createdDate": {
                        "type": "string"
                    },
                    "updatedDate": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    }
                }
            },
...
            "AddProductClientDto": {
                "type": "object",
                "properties": {
                    "productId": {
                        "type": "string"
                    },
                    "policyType": {
                        "type": "string"
                    },
                    "tpaId": {
                        "type": "string"
                    },
                    "insurerCompanyId": {
                        "type": "string"
                    }
                }
            },
...
            "AddClientListUserDto": {
                "type": "object",
                "properties": {
                    "employeeId": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "designationId": {
                        "type": "string"
                    },
                    "mailId": {
                        "type": "string"
                    },
                    "phoneNo": {
                        "type": "string"
                    }
                }
            },
...
            "ProductClientDto": {
                "type": "object",
                "properties": {
                    "pid": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "productType": {
                        "type": "string"
                    },
                    "policyType": {
                        "type": "string"
                    },
                    "tpaList": {
                        "type": "string"
                    },
                    "insurerCompany": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}